feat: Adds Globals Styles
